# api/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenClause API
  description: >
    Single entrypoint for AI agent tool-call governance.
    Validates requests, evaluates policy, manages approvals, and records audit evidence.
  version: "1.0.0"
  contact:
    name: OpenClause Team
servers:
  - url: http://localhost:8080
    description: Local gateway
  - url: http://localhost:8081
    description: Local approvals service
  - url: http://localhost:8090
    description: Local console API
security:
  - ApiKeyAuth: []
paths:
...
  /v1/toolcalls:
    post:
      operationId: submitToolCall
      summary: Submit a tool-call request for policy evaluation and execution
      tags: [Gateway]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToolCallRequest"
      responses:
        "200":
          description: Decision returned (allow, deny, or approve)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "401":
          description: Unauthorized — missing or invalid API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /v1/toolcalls/{event_id}:
    get:
      operationId: getToolCallEvent
      summary: Fetch a tool-call event by ID
      tags: [Gateway]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "404":
          description: Event not found
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
  /v1/toolcalls/{event_id}/execute:
    post:
      operationId: executeApprovedToolCall
      summary: Execute a previously approved tool-call event
      tags: [Gateway]
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Execution completed or idempotent replay returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolCallResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "404":
          description: Event not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "409":
          description: Event cannot be executed yet or does not require approval execution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIError"

  # ── Approvals ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ToolCallRequest:
      type: object
      required: [tenant_id, agent_id, tool, action, idempotency_key]
      properties:
        tenant_id:
          type: string
        agent_id:
          type: string
        tool:
          type: string
          description: "Normalized lowercase, e.g. slack"
        action:
          type: string
          description: "Normalized lowercase dotted, e.g. msg.post"
        params:
          type: object
          description: "Arbitrary JSON, max 64KB"
        resource:
          type: string
          maxLength: 2048
        risk_score:
          type: integer
          minimum: 0
          maximum: 10
          description: >
            Risk score (0-10). While optional in the schema, omitting it causes
            policy evaluation to default-deny since OPA comparisons against
            undefined values produce false. Recommended to always provide.
        risk_factors:
          type: array
          items:
            type: string
        user_id:
          type: string
        session_id:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
          maxProperties: 50
        source_ip:
          type: string
        trace_id:
          type: string
        idempotency_key:
          type: string
        requested_at:
          type: string
          format: date-time
        schema_version:
          type: string
          default: "1.0"
    ToolCallResponse:
      type: object
      properties:
        event_id:
          type: string
        decision:
          type: string
          enum: [allow, deny, approve]
        reason:
          type: string
        approval_url:
          type: string
        result:
          $ref: "#/components/schemas/ExecutionResult"
    ToolCallEnvelope:
      type: object
      properties:
        event_id:
          type: string
        request:
          $ref: "#/components/schemas/ToolCallRequest"
        decision:
          type: string
        policy_result:
          $ref: "#/components/schemas/PolicyResult"
        execution_result:
          $ref: "#/components/schemas/ExecutionResult"
        hash:
          type: string
        prev_hash:
          type: string
        received_at:
          type: string
          format: date-time
    PolicyResult:
      type: object
      properties:
        decision:
          type: string
          enum: [allow, deny, approve]
        reason:
          type: string
        requirements:
          type: object
          additionalProperties:
            type: string
        approver_group:
          type: string
        notify:
          type: array
          items:
            $ref: "#/components/schemas/PolicyNotify"
    PolicyNotify:
      type: object
      properties:
        kind:
          type: string
          enum: [webhook, slack]
        url:
          type: string
        secret_ref:
          type: string
        channel:
          type: string
    ExecutionResult:
      type: object
      properties:
        status:
          type: string
          enum: [success, error, timeout]
        output_json:
          type: object
        error:
          type: string
        duration_ms:
          type: integer

    # ── Approvals ────────────────────────────────────────────────────────
...
    APIError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        details:
          type: object